Dewiback preparations for project

+++ gip.core.datamodel/EntityACClass/ACClassConfig.cs
@@ -114,7 +114,7 @@ namespace gip.core.datamodel
         public override MsgWithDetails DeleteACObject(IACEntityObjectContext database, bool withCheck, bool softDelete = false)
         {
             MsgWithDetails msg = null;
-            if (this.ACClassPropertyRelation != null)
+            if (this.ACClassPropertyRelation != null && !softDelete)
                 this.ACClassPropertyRelation.DeleteACObject(database, withCheck);
             database.DeleteObject(this);
             return msg;
